refactor: components import types from shared types layer, not server
