+++ app/components/mainview/gmscreens/GMScreensView.tsx
@@ -3,7 +3,7 @@ import { Plus, Layers, Loader2, AlertTriangle } from 'lucide-react'
 import { useGMScreenList, useGMScreenDetail, useGMScreenMutations } from '~/hooks/useGMScreens'
 import { FloatingWindowManager, type ManagedWindow } from '~/components/mainview/FloatingWindowManager'
 import type { FloatingWindowState } from '~/components/mainview/FloatingWindow'
-import type { WindowState } from '~/server/db/models/GMScreen'
+import type { WindowState } from '~/types/gmscreen'
 import { ScreenBar } from './ScreenBar'
 import { StackCard } from './StackCard'
 import { ScreenNameDialog } from './ScreenNameDialog'
